+++ lib/magic/presenter.rb
@@ -20,7 +20,7 @@ module Magic # :nodoc:
 		], to: Base
 	end
 
-	module_function
+	module_function # TODO: extract to Magic Support
 
 	def eager_load *scopes, engine: Rails.application
 		return if Rails.application.config.eager_load
@@ -31,5 +31,12 @@ module Magic # :nodoc:
 				.map { engine.root / 'app' / _1 }
 				.select(&:exist?)
 				.each { Rails.autoloaders.main.eager_load_dir _1 }
+	end
+
+	def each_engine(&)
+		Rails.application
+				.then { [ _1, *_1.railties ] }
+				.grep(Rails::Engine)
+				.each(&)
 	end
 end